// API/src/types/player.ts
export interface IPlayer {
  name: string
  height: number | null
  width: number | null
  college: string | null
  born: number | null
  birth_city:string | null
  birth_state: string | null
  year_start:number | null
  year_end:number | null
  position:string | null
  weight: number | null
  birth_date: Date | null
}